Time: 0 ms (100%), Space: 41.9 MB (72.97%) - LeetHub

+++ 0367-valid-perfect-square/0367-valid-perfect-square.java
@@ -5,7 +5,7 @@ class Solution {
         while(s<=e)
         {
             int mid = (s+e)/2;
-            if(mid*mid==num)
+            if((long)mid*mid==num)
             {
                 return true;
             }
